# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_shendrong_fullstack-go_server_internal_model.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.UserResponse'
    type: object
  github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
  github_com_shendrong_fullstack-go_server_internal_model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_shendrong_fullstack-go_server_internal_model.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      full_name:
        maxLength: 255
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  github_com_shendrong_fullstack-go_server_internal_model.UpdateProfileRequest:
    properties:
      avatar_url:
        maxLength: 2048
        type: string
      full_name:
        maxLength: 255
        minLength: 1
        type: string
    type: object
  github_com_shendrong_fullstack-go_server_internal_model.UploadResponse:
    properties:
      filename:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
  github_com_shendrong_fullstack-go_server_internal_model.UserResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Production-grade REST API with JWT authentication, file upload, and
    PostgreSQL.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Fullstack Go API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, returns JWT token
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
    patch:
      consumes:
      - application/json
      description: Allows updating full_name and/or avatar_url of the currently authenticated
        user
      parameters:
      - description: Profile update fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /health:
    get:
      description: Returns server and database connection status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - system
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file (image or document). Returns the file URL. Max size
        10MB.
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.UploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_shendrong_fullstack-go_server_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a file
      tags:
      - files
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"